[JVM] NDArray from & to different dtypes

+++ jvm/core/src/main/scala/ml/dmlc/tvm/LibInfo.scala
@@ -45,7 +45,7 @@ private[tvm] class LibInfo {
                             ctx: TVMContext,
                             refHandle: RefTVMArrayHandle): Int
   @native def tvmArrayGetShape(handle: TVMArrayHandle, shape: ArrayBuffer[Long]): Int
-  @native def tvmArrayCopyFromJArray(fromRaw: Array[Float],
+  @native def tvmArrayCopyFromJArray(fromRaw: Array[Byte],
                                      from: TVMArrayHandle,
                                      to: TVMArrayHandle): Int
   @native def tvmArrayCopyToJArray(from: TVMArrayHandle, to: Array[Byte]): Int
